+++ src/module/applications/ux/enrichers/lookup.mjs
@@ -42,8 +42,10 @@ export async function enricher(match, options) {
 
   /** @type {string | number} */
   let value;
-  if (parsedConfig.evaluate) value = ds.utils.evaluateFormula(parsedConfig.formula, data, { contextName: "lookup" });
-  else {
+  if (parsedConfig.evaluate) {
+    const replacedFormula = Roll.replaceFormulaData(parsedConfig.formula, data);
+    value = (replacedFormula.includes("@")) ? fallback : ds.utils.evaluateFormula(replacedFormula, data, { contextName: "lookup" });
+  } else {
     value = foundry.utils.getProperty(data, parsedConfig.formula.substring(1)) ?? fallback;
 
     switch (parsedConfig.style) {
@@ -64,6 +66,7 @@ export async function enricher(match, options) {
   if (!value && (options.documents === false)) return null;
   else if (!value) span.classList.add("not-found");
   span.innerText = value ?? parsedConfig.formula;
+  span.dataset.tooltip = (value === fallback) ? parsedConfig.formula : fallback;
   return span;
 }