+++ src/module/data/helpers.mjs
@@ -16,11 +16,9 @@ export const requiredInteger = ({ initial = 0, min = 0, max, label } = {}) => ne
 /**
  * Constructs a string field for use inside of a SetField.
  * @param {object} [options] Options to forward to the field.
- * @param {Record<string, string>} [options.choices] CONST-derived choices for the field.
- * @param {Function} [options.validate] A validator function for field values.
  * @returns A string field that is always truthy.
  */
-export const setOptions = ({ choices, validate } = {}) => new StringField({ required: true, blank: false, choices, validate });
+export const setOptions = (options) => new StringField({ required: true, blank: false, ...options });
 
 /* -------------------------------------------------- */
 
@@ -47,4 +45,19 @@ export const damageTypes = (inner, { all = false } = {}) => {
   }, schema);
 
   return new SchemaField(schema);
+};
+
+/* -------------------------------------------------- */
+
+/**
+ * Validates if a DSID fulfills our parameters.
+ * @param {string} dsid
+ */
+export const validateDSID = (dsid) => {
+  // Valid slug
+  const slug = dsid.slugify({ strict: true });
+  // multi-slugging can continue to mutate, this checks if the mutation is *only* from the middle mutation.
+  const reSlug = dsid.replace(/[\s-]+/g, "-");
+
+  return [dsid, reSlug].includes(slug);
 };