+++ src/module/data/item/ability.mjs
@@ -1,11 +1,11 @@
 import { systemPath } from "../../constants.mjs";
-import { DrawSteelActiveEffect, DrawSteelActor, DrawSteelChatMessage } from "../../documents/_module.mjs";
-import { DamageRoll, PowerRoll } from "../../rolls/_module.mjs";
+import { DrawSteelActiveEffect, DrawSteelChatMessage } from "../../documents/_module.mjs";
+import { PowerRoll } from "../../rolls/_module.mjs";
 import FormulaField from "../fields/formula-field.mjs";
-import { setOptions } from "../helpers.mjs";
+import { setOptions, validateDSID } from "../helpers.mjs";
 import enrichHTML from "../../utils/enrich-html.mjs";
 import DamagePowerRollEffect from "../pseudo-documents/power-roll-effects/damage-effect.mjs";
-import BaseItemModel from "./base.mjs";
+import BaseItemModel from "./base-item.mjs";
 
 /**
  * @import { DocumentHTMLEmbedConfig, EnrichmentOptions } from "@client/applications/ux/text-editor.mjs";
@@ -44,6 +44,16 @@ export default class AbilityModel extends BaseItemModel {
 
     // Items don't have descriptions
     delete schema.description;
+
+    // Can be expanded over time for automation
+    schema.prerequisites = new fields.SchemaField({
+      value: new fields.StringField({ required: true }),
+      dsid: new fields.SetField(setOptions({
+        validate: validateDSID,
+        validationError: game.i18n.localize("DRAW_STEEL.SOURCE.InvalidDSID"),
+      })),
+      level: new fields.NumberField({ required: true, integer: true, positive: true }),
+    });
 
     schema.story = new fields.StringField({ required: true });
     schema.keywords = new fields.SetField(setOptions());
@@ -335,10 +345,10 @@ export default class AbilityModel extends BaseItemModel {
 
     context.characteristics = Object.entries(ds.CONFIG.characteristics).map(([value, { label }]) => ({ value, label }));
 
-    context.powerRollEffects = Object.fromEntries([1, 2, 3].map(tier => [
-      `tier${tier}`,
-      this.powerRollText(tier),
-    ]));
+    context.powerRollEffects = {};
+    for (const tier of [1, 2, 3]) {
+      context.powerRollEffects[`tier${tier}`] = await this.powerRollText(tier);
+    }
     context.powerRolls = this.power.effects.size > 0;
 
     context.powerRollBonus = this.power.roll.formula;
@@ -367,9 +377,9 @@ export default class AbilityModel extends BaseItemModel {
   /**
    * Produces the power roll text for a given tier.
    * @param {1 | 2 | 3} tier
-   * @returns {string}
+   * @returns {Promise<string>} An HTML string.
    */
-  powerRollText(tier) {
+  async powerRollText(tier) {
     return this.power.effects.sortedContents.map(effect => effect.toText(tier)).filter(_ => _).join("; ");
   }