+++ src/module/data/item/feature.mjs
@@ -1,3 +1,5 @@
+import { systemPath } from "../../constants.mjs";
+import { setOptions, validateDSID } from "../helpers.mjs";
 import AdvancementModel from "./advancement.mjs";
 
 /**
@@ -10,8 +12,7 @@ export default class FeatureModel extends AdvancementModel {
       ...super.metadata,
       type: "feature",
       packOnly: false,
-      // Not currently in use
-      // detailsPartial: [systemPath("templates/sheets/item/partials/feature.hbs")],
+      detailsPartial: [systemPath("templates/sheets/item/partials/feature.hbs")],
     };
   }
 
@@ -27,7 +28,15 @@ export default class FeatureModel extends AdvancementModel {
     const fields = foundry.data.fields;
     const schema = super.defineSchema();
 
-    // TODO: https://github.com/MetaMorphic-Digital/draw-steel/issues/839 will re-add content here.
+    // Can be expanded over time for automation
+    schema.prerequisites = new fields.SchemaField({
+      value: new fields.StringField({ required: true }),
+      dsid: new fields.SetField(setOptions({
+        validate: validateDSID,
+        validationError: game.i18n.localize("DRAW_STEEL.SOURCE.InvalidDSID"),
+      })),
+      level: new fields.NumberField({ required: true, integer: true, positive: true }),
+    });
 
     return schema;
   }